docId as long instead of ulong

+++ src/Sir.Store/DocIndexReader.cs
@@ -30,9 +30,9 @@ namespace Sir.Store
         /// </summary>
         /// <param name="docId">Document ID</param>
         /// <returns>The offset and length of a document's key_id/value_id map</returns>
-        public (long offset, int length) Read(ulong docId)
+        public (long offset, int length) Read(long docId)
         {
-            var offs = (long)docId * _blockSize;
+            var offs = docId * _blockSize;
 
             _stream.Seek(offs, SeekOrigin.Begin);
 
@@ -52,7 +52,7 @@ namespace Sir.Store
         /// </summary>
         /// <param name="docId">Document ID</param>
         /// <returns>The offset and length of a document's key_id/value_id map</returns>
-        public async Task<(long offset, int length)> ReadAsync(ulong docId)
+        public async Task<(long offset, int length)> ReadAsync(long docId)
         {
             var offs = (long) docId * _blockSize;
